Corrected problem when selecting cancel in path selection dialog

+++ configurationGui.py
@@ -86,11 +86,12 @@ class configurationGUI:
 	def defaultOriginFileChooser(self):
 		global windowDefaultOrigin
 		
-		windowDefaultOrigin = tkFileDialog.askdirectory(initialdir = windowDefaultOrigin) + "/"
-		windowDefaultOrigin = windowDefaultOrigin.encode("utf-8")
-		if windowDefaultOrigin != "":
-			self.textDefaultOriginPath.delete('1.0', END)
-			self.textDefaultOriginPath.insert(END, windowDefaultOrigin)
+		resultPath = tkFileDialog.askdirectory(initialdir = windowDefaultOrigin) + "/"
+		if resultPath != "/":
+			windowDefaultOrigin = resultPath.encode("utf-8")
+			if windowDefaultOrigin != "":
+				self.textDefaultOriginPath.delete('1.0', END)
+				self.textDefaultOriginPath.insert(END, windowDefaultOrigin)
 	#Finished originFileChooser
 	
 	def accept(self):
